requirements.txt
streamlit>=1.30.0
openai>=1.6.1
pandas>=2.1.0
numpy>=1.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
scikit-learn>=1.3.0
